backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.26.0
pillow>=10.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
tensorflow>=2.12.0
kaggle>=1.6.14
